+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/mapper/graphtransfo/BufferProperties.java
@@ -47,7 +47,7 @@ public class BufferProperties {
 
   private final String dataTypeName;
   private final String destInputPortID;
-  private final int    size;
+  private final long   size;
   private final String sourceOutputPortID;
 
   /**
@@ -61,7 +61,7 @@ public class BufferProperties {
    *          the size
    */
   public BufferProperties(final String dataTypeName, final String sourceOutputPortID, final String destInputPortID,
-      final int size) {
+      final long size) {
     super();
     this.dataTypeName = dataTypeName;
     this.destInputPortID = destInputPortID;
@@ -77,7 +77,7 @@ public class BufferProperties {
     return this.destInputPortID;
   }
 
-  public int getSize() {
+  public long getSize() {
     return this.size;
   }
